requirements.txt
requests>=2.28.0
python-dotenv
